fix(dashboard): allow change certain types of custom fields
